fix markdown rendering
