+++ views/blank.tpl
@@ -1,0 +1,1 @@
+{{.Text}}
